+++ client/assets/entries/app.tsx
@@ -1,23 +1,15 @@
 // Can't use import React from "react",
 // because it is not exported as default.
-import * as React from "react";
-import * as ReactDOM from "react-dom";
-
-import { Provider } from "react-redux";
-import Store from "scripts/redux/store";
+import * as React from 'react';
+import * as ReactDOM from 'react-dom';
 
 // Import container.
-import App from "scripts/redux/containers/app";
+import App from 'scripts/app';
 
 // Import bootstrap.
 // import "bootstrap/dist/css/bootstrap.min.css";
 
 // Import custom stylesheets.
-import "stylesheets/entries/app.scss";
+import 'stylesheets/entries/app.scss';
 
-ReactDOM.render(
-    <Provider store={ Store }>
-        <App />
-    </Provider>,
-    document.getElementById("app")
-);
+ReactDOM.render(<App />, document.getElementById('app'));
